code-splitting Matteo changed editor config, renamed some components and added destructuring
