+++ .editorconfig
@@ -1,15 +1,15 @@
 root = true
 
 [*]
-indent_style = space
+indent_style = tab
 end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
 
 [{package.json,.*rc,*.yml}]
-indent_style = space
-indent_size = 2
+indent_style = tab
+indent_size = 4
 
 [*.md]
 trim_trailing_whitespace = false
