+++ src/components/utils/index.js
@@ -1,0 +1,2 @@
+export If from './if';
+export SplitCode from './splitCode';
